% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nic_elecciones}
\alias{nic_elecciones}
\title{Elecciones presidenciales en Nicaragua}
\format{
A data frame that contains 72 rows and 35 variables:
\describe{
\item{year}{year}
\item{departamento}{Departamento}
\item{municipios}{Municipios en Nicaragua}
\item{inscritos}{Inscritos}
\item{jrv}{Junta Receptora de Votos}
\item{total_votos}{Totol de votos}
\item{total_porcentaje}{Porcentaje de votos}
\item{nulos}{Total de votos nulos}
\item{nulos_porcentaje}{Porcentaje de votos nulos}
\item{votos_validos}{Total de votos válidos}
\item{validos_porcentaje}{Porcentaje de votos válidos}
\item{uno}{PP UNO}
\item{psoc}{PP PSOC}
\item{pliun}{PP PLIUN}
\item{prt}{PP PRT}
\item{fsln}{PP FSLN}
\item{map_ml}{PP MAP ML}
\item{psc}{PP PSC}
\item{puca}{PP PUCA}
\item{pcdn}{PP PCDN}
\item{mur}{PP MUR}
\item{al}{PP AL}
\item{pcn}{PP PCN}
\item{ccn}{PP CCN}
\item{pronal}{PP PRONAL}
\item{mrs}{PP MRS}
\item{prn}{PP PRN}
\item{anc}{PP ANC}
\item{u}{PP U}
\item{pli}{PP PLI}
\item{uno_96}{PP UNO 96}
\item{plc}{PP PLC}
\item{ac}{PP AC}
\item{aln}{PP ALN}
\item{otros}{PP OTROS}
}
}
\usage{
nic_elecciones
}
\description{
A dataset containing the results of the presidenciales in Nicaragua
}
\keyword{datasets}
